Restore unaltered contributor list and give details on this version
The contributor list is an invariant part of the original book, meaning
that we are not supposed to alter it. Yet, it should be clear that this
version of the book is not the original, that the contributor list
applies to the original, and that the book is available on github.

Fixes #7