Make while/for loop examples more consistent
The `while` example sums integers from 1 up to, and including, `n`. The simplified `for i in range(n+1)` that we give as the better alternative has the same outcome, but is slightly different in that the loop is traversed also for `i = 0`.

Since we say the behavior is exactly the same it's best to have the `range` alternative also skip the zero case. This was, rightly, confusing a student who was tracing the behavior of each loop cycle-by-cycle.